Revert "Detail Activity Dagger Modules"
