+++ app/src/main/java/com/example/nvdovin/weatherapp/domain/utils/design/ImageUtils.java
@@ -5,6 +5,8 @@ import android.os.Build;
 import android.text.Html;
 import android.widget.TextView;
 
+import com.example.nvdovin.weatherapp.domain.utils.mapper.WeatherCodesMap;
+
 public class ImageUtils {
 
     private static final String FONTS_LOCATION = "fonts/weathericons-regular-webfont.ttf";
