+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/application/dagger/DatabaseModule.java
@@ -6,7 +6,7 @@ import com.example.nvdovin.weatherapp.data.dao.DaoMaster;
 import com.example.nvdovin.weatherapp.data.dao.DaoSession;
 import com.example.nvdovin.weatherapp.domain.service.CityService;
 import com.example.nvdovin.weatherapp.domain.service.WeatherDataService;
-import com.example.nvdovin.weatherapp.domain.utils.mapper.DataMapper;
+import com.example.nvdovin.weatherapp.domain.utils.updater.DataMapper;
 
 import dagger.Module;
 import dagger.Provides;
