+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/application/dagger/ExecutorModule.java
@@ -7,7 +7,7 @@ import com.example.nvdovin.weatherapp.data.network.api.WeatherApi;
 import com.example.nvdovin.weatherapp.domain.service.CityService;
 import com.example.nvdovin.weatherapp.domain.service.WeatherDataService;
 import com.example.nvdovin.weatherapp.domain.utils.executor.Executor;
-import com.example.nvdovin.weatherapp.domain.utils.mapper.DataMapper;
+import com.example.nvdovin.weatherapp.domain.utils.updater.DataMapper;
 
 import dagger.Module;
 import dagger.Provides;
