--- app/src/main/java/com/example/nvdovin/weatherapp/presentation/details/core/ViewCallback.java
@@ -1,6 +1,0 @@
-package com.example.nvdovin.weatherapp.presentation.details.core;
-
-
-public interface ViewCallback {
-    void historyClickHandler();
-}
