--- app/src/main/java/com/example/nvdovin/weatherapp/presentation/details/dagger/DetailScope.java
@@ -1,12 +1,0 @@
-package com.example.nvdovin.weatherapp.presentation.details.dagger;
-
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-import javax.inject.Scope;
-
-@Scope
-@Retention(RetentionPolicy.CLASS)
-public @interface DetailScope {
-}
