+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/history/grid/GridHistoryPresenter.java
@@ -44,4 +44,5 @@ public class GridHistoryPresenter {
     public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
+
 }
