+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/history/grid/GridHistoryView.java
@@ -45,6 +45,7 @@ public class GridHistoryView {
 
         weatherGridView.setAdapter(gridHistoryAdapter);
 
+
     }
 
     public View getView() {
