+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/history/grid/dagger/GridHistoryFragmentComponent.java
@@ -5,6 +5,10 @@ import com.example.nvdovin.weatherapp.presentation.history.grid.GridHistoryFragm
 
 import dagger.Component;
 
+/**
+ * Created by nvdovin on 11/24/2016.
+ */
+
 @Component(modules = GridHistoryFragmentModule.class, dependencies = AppComponent.class)
 @GridHistoryFragmentScope
 public interface GridHistoryFragmentComponent {
