+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/history/grid/dagger/GridHistoryFragmentModule.java
@@ -1,6 +1,7 @@
 package com.example.nvdovin.weatherapp.presentation.history.grid.dagger;
 
-import com.example.nvdovin.weatherapp.domain.utils.mapper.DataMapper;
+import com.example.nvdovin.weatherapp.domain.service.CityService;
+import com.example.nvdovin.weatherapp.domain.service.WeatherDataService;
 import com.example.nvdovin.weatherapp.presentation.history.grid.GridHistoryFragment;
 import com.example.nvdovin.weatherapp.presentation.history.grid.GridHistoryPresenter;
 import com.example.nvdovin.weatherapp.presentation.history.grid.GridHistoryView;
@@ -20,20 +21,26 @@ public class GridHistoryFragmentModule {
 
     @Provides
     @GridHistoryFragmentScope
+    GridHistoryFragment provideGridHistoryFragment() {
+        return gridHistoryFragment;
+    }
+
+    @Provides
+    @GridHistoryFragmentScope
     GridHistoryAdapter provideGridHistoryAdapter() {
         return new GridHistoryAdapter(gridHistoryFragment.getActivity());
     }
 
     @Provides
     @GridHistoryFragmentScope
-    GridHistoryView provideGridHistoryView(GridHistoryAdapter gridHistoryAdapter) {
+    GridHistoryView provideGridHistoryView(GridHistoryFragment gridHistoryFragment, GridHistoryAdapter gridHistoryAdapter) {
         return new GridHistoryView(gridHistoryFragment.getActivity(), gridHistoryAdapter);
     }
 
     @Provides
     @GridHistoryFragmentScope
-    GridHistoryPresenter provideGridHistoryPresenter(DataMapper dataMapper, GridHistoryView gridHistoryView) {
-        return new GridHistoryPresenter(dataMapper, gridHistoryView);
+    GridHistoryPresenter provideGridHistoryPresenter(CityService cityService, WeatherDataService weatherDataService, GridHistoryView gridHistoryView) {
+        return new GridHistoryPresenter(cityService, weatherDataService, gridHistoryView);
     }
 
 }
