+++ app/src/main/java/com/example/nvdovin/weatherapp/presentation/main/weather/forecast/dagger/ForecastFragmentModule.java
@@ -1,11 +1,10 @@
 package com.example.nvdovin.weatherapp.presentation.main.weather.forecast.dagger;
-
 import com.example.nvdovin.weatherapp.data.SortQueryBuilder;
 import com.example.nvdovin.weatherapp.domain.service.CityService;
 import com.example.nvdovin.weatherapp.domain.utils.executor.DefaultThreadPoolExecutor;
 import com.example.nvdovin.weatherapp.domain.utils.executor.Executor;
 import com.example.nvdovin.weatherapp.domain.utils.sharedpreferences.SharedPrefs;
-import com.example.nvdovin.weatherapp.domain.utils.mapper.DataMapper;
+import com.example.nvdovin.weatherapp.domain.utils.updater.DataMapper;
 import com.example.nvdovin.weatherapp.presentation.main.weather.forecast.ForecastFragment;
 import com.example.nvdovin.weatherapp.presentation.main.weather.forecast.ForecastPresenter;
 import com.example.nvdovin.weatherapp.presentation.main.weather.forecast.ForecastView;
