+++ app/src/main/res/layout/detail_main_layout.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/rain_bckg"
     android:fitsSystemWindows="true"
     app:layout_collapseMode="parallax">
-
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/main.appbar"
@@ -52,5 +51,5 @@
         android:layout_marginBottom="@dimen/standart_margin"
         android:clipToPadding="false"
         app:layout_behavior="@string/appbar_scrolling_view_behavior" />
-</android.support.design.widget.CoordinatorLayout>
 
+</android.support.design.widget.CoordinatorLayout>
